// src/app/hairy-paws/interfaces/pet/pet-interface.ts
export interface PetInterface {
  id: string
  name: string
  type: string
  breed: string
  age: number
  gender: string
  description: string
  ownerId: string
  images: string[]
  status: string
  availableForAdoption: boolean
  weight: string
  healthDetails: string
  vaccinated: boolean
  sterilized: boolean
}